requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
numpy
python-multipart
